+++ tests/environments/local_flip_test_cases.py
@@ -11,7 +11,7 @@ from .local_set_test_cases import (
 )
 
 DTYPE_STATE = np.uint8
-DTYPE_ACTION = np.uint8
+DTYPE_ACTION = np.int32
 
 TEST_CASES_CONSTRUCTOR = insert(
     remove_at([tc for tc in lstc_constructor if tc[5] == 2], 5), 4, False
