+++ tests/environments/local_set_test_cases.py
@@ -9,7 +9,7 @@ from ..utils import batchify, replace
 from .global_set_test_cases import TEST_CASES_CONSTRUCTOR as gstc_constructor
 
 DTYPE_STATE = np.uint8
-DTYPE_ACTION = np.int8
+DTYPE_ACTION = np.int32
 
 TEST_CASES_CONSTRUCTOR = deepcopy(gstc_constructor)
